requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
tavily-python
python-dotenv
groq
reportlab
pymongo
pydantic
python-jose[cryptography]
argon2-cffi
pyjwt
email-validator